# openapi.yaml
openapi: 3.0.3
info:
  title: Reconciliation API
  version: RECONCILIATION_VERSION
paths:
...
  /alerts:
    get:
      summary: List alerts
      operationId: listAlerts
      tags: [reconciliation.v1]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QueryBuilder' }
      responses:
        '200': { $ref: '#/components/responses/Alerts' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:read]
  /alerts/{alertID}:
    parameters: [{ $ref: '#/components/parameters/AlertID' }]
    get:
      summary: Get an alert
      operationId: getAlert
      tags: [reconciliation.v1]
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:read]
  /alerts/{alertID}/events:
    parameters: [{ $ref: '#/components/parameters/AlertID' }]
    get:
      summary: List alert events (append-only timeline)
      operationId: listAlertEvents
      description: |
        Returns a page of the events recorded for this alert — every evaluation
        that touched it plus every manual transition. The list is append-only;
        events are never modified or deleted. Ordered most-recent-first and
        cursor-paginated: a long-lived alert's timeline is unbounded (one row per
        failing evaluation), so callers must page through it.
      tags: [reconciliation.v1]
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200': { $ref: '#/components/responses/AlertEvents' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:read]
  /alerts/{alertID}/ack:
    post:
      summary: Acknowledge an alert
      operationId: ackAlert
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/AlertID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AckAlertRequest' }
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
  /alerts/{alertID}/resolve:
    post:
      summary: Resolve an alert (fixed_by_booking)
      operationId: resolveAlert
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/AlertID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveAlertRequest' }
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
  /alerts/{alertID}/accept:
    post:
      summary: Accept an alert (accepted_by_business)
      operationId: acceptAlert
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/AlertID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AcceptAlertRequest' }
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
  /alerts/{alertID}/snooze:
    post:
      summary: Snooze an alert's notifications until a future instant
      description: |
        Mutes the alert's webhook notifications until `until`. The alert keeps
        failing, keeps its status, and keeps counting against period-green —
        only its notifications are suppressed, even if the discrepancy changes.
        The first failing evaluation at or after `until` clears the snooze and
        notifies once. Re-snoozing overwrites the window. Rejects RESOLVED
        alerts and a non-future `until`.
      operationId: snoozeAlert
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/AlertID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SnoozeAlertRequest' }
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
  /alerts/{alertID}/unsnooze:
    post:
      summary: Lift a snooze early
      description: |
        Clears an active snooze before its window elapses. Idempotent —
        unsnoozing an alert that is not snoozed returns it unchanged.
      operationId: unsnoozeAlert
      tags: [reconciliation.v1]
      parameters: [{ $ref: '#/components/parameters/AlertID' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UnsnoozeAlertRequest' }
      responses:
        '200': { $ref: '#/components/responses/Alert' }
        default: { $ref: '#/components/responses/ErrorResponse' }
      security:
        - Authorization: [reconciliation:write]
...
components:
  parameters:
    PageSize:
      name: pageSize
      in: query
      description: |
        The maximum number of results to return per page.
      example: 100
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 1000
    Cursor:
      name: cursor
      in: query
      description: |
        Parameter used in pagination requests. Maximum page size is set to 15.
        Set to the value of next for the next page of results.
        Set to the value of previous for the previous page of results.
        No other parameters can be set when this parameter is set.
      schema:
        type: string
        example: aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ==
...
    AlertID:
      name: alertID
      in: path
      required: true
      description: The alert ID.
      schema: { type: string, format: uuid }
  responses:
...
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Alert:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlertResponse' }
    Alerts:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlertsCursorResponse' }
    AlertEvents:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AlertEventsCursorResponse' }
...
  schemas:
    QueryBuilder:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          example: VALIDATION
        errorMessage:
          type: string
        details:
          type: string

    # --- V1 Ledger Clarity schemas -----------------------------------------
    Severity:
      type: string
      enum: [info, low, medium, high, critical]
...
    Ack:
      type: object
      required: [by, at]
      properties:
        by: { type: string, example: "ops@buildr.com" }
        at: { type: string, format: date-time }
        note: { type: string }
    Resolution:
      type: object
      required: [kind, by, at]
      properties:
        kind: { type: string, enum: [auto, fixed_by_booking, accepted_by_business] }
        by: { type: string }
        at: { type: string, format: date-time }
        note: { type: string }
        transactionRefs:
          type: array
          items: { type: string }
        evidenceSnapshot:
          type: object
          additionalProperties: true
        expiresAt: { type: string, format: date-time }
    Snooze:
      type: object
      description: |
        A time-boxed, operator-initiated mute of an alert's notifications. While
        `until` is in the future the alert keeps failing and keeps counting
        against period-green — only its notifications are suppressed.
      required: [until, by, at]
      properties:
        until: { type: string, format: date-time }
        by: { type: string, example: "ops@buildr.com" }
        at: { type: string, format: date-time }
        note: { type: string }
...
    Alert:
      type: object
      description: |
        An Alert is the stable, dedup'd record of a failing fingerprint within a
        reconciliation period. There is exactly one Alert per
        (ruleID, fingerprint, periodID) — within a period, re-opens after resolve
        flip the status back to OPEN in place; the same fingerprint failing in a
        new period is a separate Alert. For a `continuous`-cadence rule there is
        one unbounded period, so it behaves as one Alert per (ruleID, fingerprint).
        The full history of transitions lives in alert_event
        (see /alerts/{alertID}/events).
      required: [id, ruleID, fingerprint, periodID, status, severity, firstSeenAt, lastSeenAt, occurrenceCount, lastEvaluationID, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ruleID: { type: string, format: uuid }
        fingerprint: { type: string, example: "asset:USD/2" }
        periodID: { type: string, example: "2026-03", description: "Reconciliation period this alert is scoped to ('continuous' for live-monitoring rules)." }
        status: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        severity: { $ref: '#/components/schemas/Severity' }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        occurrenceCount: { type: integer, format: int64, description: "Count of FAIL events on this alert across reopen cycles within its period (lifetime, for a continuous-cadence rule)." }
        lastEvaluationID: { type: string, format: uuid }
        evidence:
          type: object
          additionalProperties: true
        ack: { $ref: '#/components/schemas/Ack' }
        resolution: { $ref: '#/components/schemas/Resolution' }
        snooze: { $ref: '#/components/schemas/Snooze' }
        labels:
          type: object
          additionalProperties: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AlertResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: '#/components/schemas/Alert' }
    AlertsCursorResponse:
      type: object
      required: [cursor]
      properties:
        cursor:
          type: object
          required: [pageSize, hasMore, data]
          properties:
            pageSize: { type: integer, format: int64 }
            hasMore: { type: boolean }
            previous: { type: string }
            next: { type: string }
            data:
              type: array
              items: { $ref: '#/components/schemas/Alert' }
    AlertEvent:
      type: object
      description: |
        One row of the alert's append-only event log. Captures every
        evaluation that touched the alert plus every manual transition
        (ack / resolve / accept / snooze / unsnooze). `prevStatus` is null
        only for the inaugural event of an alert. `isReopen` is a derived
        predicate — true when a FAIL event lands on a previously-RESOLVED alert.
        `notify` is the notification decision recorded for the row: false marks
        a transition kept for audit but deliberately not published (a repeated
        identical fail, or a fail muted by an active snooze).
      required: [id, alertID, type, newStatus, at, isReopen, notify]
      properties:
        id: { type: string, format: uuid }
        alertID: { type: string, format: uuid }
        evaluationID: { type: string, format: uuid, nullable: true }
        type: { type: string, enum: [fail, pass, ack, resolve, accept, snooze, unsnooze] }
        prevStatus: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED], nullable: true }
        newStatus: { type: string, enum: [OPEN, ACKNOWLEDGED, RESOLVED] }
        payload:
          description: Event-type-specific payload (evidence for fail, resolution for resolve/accept, ack object for ack, snooze object for snooze).
          type: object
          additionalProperties: true
        at: { type: string, format: date-time }
        isReopen: { type: boolean }
        notify: { type: boolean, description: "Whether this transition was published to the message bus. false = recorded for audit but suppressed." }
    AlertEventsCursorResponse:
      type: object
      required: [cursor]
      properties:
        cursor:
          type: object
          required: [pageSize, hasMore, data]
          properties:
            pageSize: { type: integer, format: int64 }
            hasMore: { type: boolean }
            previous: { type: string }
            next: { type: string }
            data:
              type: array
              items: { $ref: '#/components/schemas/AlertEvent' }
...
    AckAlertRequest:
      type: object
      required: [by]
      properties:
        by: { type: string, example: "ops@buildr.com" }
        note: { type: string }
    ResolveAlertRequest:
      type: object
      required: [by]
      description: |
        Mark an alert resolved. When `transactionRefs` is non-empty the
        resolution kind is recorded as `fixed_by_booking`.
      properties:
        by: { type: string }
        note: { type: string }
        transactionRefs:
          type: array
          items: { type: string }
    AcceptAlertRequest:
      type: object
      required: [by, note]
      description: Business acceptance. `note` is required and the evidence is snapshotted onto the resolution.
      properties:
        by: { type: string }
        note: { type: string, minLength: 1 }
        expiresAt: { type: string, format: date-time }
    SnoozeAlertRequest:
      type: object
      required: [by, until]
      description: Mute an alert's notifications until `until` (which must be in the future).
      properties:
        by: { type: string, example: "ops@buildr.com" }
        until: { type: string, format: date-time }
        note: { type: string }
    UnsnoozeAlertRequest:
      type: object
      required: [by]
      properties:
        by: { type: string, example: "ops@buildr.com" }

    # --- V2 named multi-source schemas ------------------------------------
...